feat: centralize ptosc option validation
